Copy texture via blit
